+++ Source/WTF/wtf/MetaAllocator.h
@@ -138,7 +138,7 @@ private:
 
         size_t sizeInBytes()
         {
-            return m_end.untaggedPtr<size_t>() - m_start.untaggedPtr<size_t>();
+            return (size_t)m_end.untaggedPtr<uintptr_t>() - (size_t)m_start.untaggedPtr<uintptr_t>();
         }
 
         size_t key()
